rename project + install deps
