+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.invertase.rnfirebasestarter",
+    package = "com.traveled.traveled",
 )
 
 android_resource(
     name = "res",
-    package = "com.invertase.rnfirebasestarter",
+    package = "com.traveled.traveled",
     res = "src/main/res",
 )
