+++ android/app/build.gradle
@@ -103,7 +103,7 @@ android {
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "com.invertase.rnfirebasestarter"
+        applicationId "com.traveled.traveled"
         minSdkVersion 16
         targetSdkVersion 22
         versionCode 1
