+++ android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.invertase.rnfirebasestarter"
+    package="com.traveled.traveled"
     android:versionCode="1"
     android:versionName="1.0">
