+++ android/app/src/main/java/com/traveled/traveled/MainActivity.java
@@ -1,4 +1,4 @@
-package com.invertase.rnfirebasestarter;
+package com.traveled.traveled;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "RNFirebaseStarter";
+        return "Traveled";
     }
 }
