+++ android/app/src/main/java/com/traveled/traveled/MainApplication.java
@@ -1,4 +1,4 @@
-package com.invertase.rnfirebasestarter;
+package com.traveled.traveled;
 
 import android.app.Application;
