+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'RNFirebaseStarter'
+rootProject.name = 'Traveled'
 include ':react-native-firebase'
 project(':react-native-firebase').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-firebase/android')
