+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "RNFirebaseStarter",
-  "displayName": "RNFirebaseStarter"
+  "name": "Traveled",
+  "displayName": "Traveled"
 }
