+++ ios/Traveled/Images.xcassets/AppIcon.appiconset/Contents.json
