+++ ios/Traveled/Info.plist
