+++ ios/TraveledTests/TraveledTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface RNFirebaseStarterTests : XCTestCase
+@interface TraveledTests : XCTestCase
 
 @end
 
-@implementation RNFirebaseStarterTests
+@implementation TraveledTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
