+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "RNFirebaseStarter",
+  "name": "Traveled",
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
